Moved model to dedicated backend
