+++ lifter/models.py
@@ -1,16 +1,8 @@
 
 from . import query
+from .backends.base import BaseModelMeta, BaseModel
 
-class BaseModelMeta(type):
-    def __getattr__(cls, key):
-        return getattr(query.Path(), key)
-
-class BaseModel(object):
-    __metaclass__ = BaseModelMeta
-
-    @classmethod
-    def load(cls, values):
-        return query.QuerySet(values)
 
 def Model(name):
-    return BaseModelMeta(name, (BaseModel,), {})
+    from .backends import python
+    return BaseModelMeta(name, (python.PythonModel,), {})
